+++ src/main/java/com/ureca/snac/infra/exception/TossInvalidApiKeyException.java
@@ -1,11 +1,12 @@
-package com.ureca.snac.payment.exception;
+package com.ureca.snac.infra.exception;
 
 import com.ureca.snac.common.exception.ExternalApiException;
 
-import static com.ureca.snac.common.BaseCode.TOSS_INVALID_API_KEY;
+import static com.ureca.snac.common.BaseCode.PAYMENT_GATEWAY_CONFIG_ERROR;
 
 public class TossInvalidApiKeyException extends ExternalApiException {
+
     public TossInvalidApiKeyException() {
-        super(TOSS_INVALID_API_KEY);
+        super(PAYMENT_GATEWAY_CONFIG_ERROR);
     }
 }
