+++ src/main/java/com/ureca/snac/infra/exception/TossInvalidCardInfoException.java
@@ -1,11 +1,12 @@
-package com.ureca.snac.payment.exception;
+package com.ureca.snac.infra.exception;
 
 import com.ureca.snac.common.exception.BusinessException;
 
-import static com.ureca.snac.common.BaseCode.TOSS_INVALID_CARD_INFO;
+import static com.ureca.snac.common.BaseCode.GATEWAY_INVALID_CARD_INFO;
 
 public class TossInvalidCardInfoException extends BusinessException {
+
     public TossInvalidCardInfoException() {
-        super(TOSS_INVALID_CARD_INFO);
+        super(GATEWAY_INVALID_CARD_INFO);
     }
 }
