+++ src/main/java/com/ureca/snac/infra/exception/TossNotEnoughBalanceException.java
@@ -1,11 +1,12 @@
-package com.ureca.snac.payment.exception;
+package com.ureca.snac.infra.exception;
 
 import com.ureca.snac.common.exception.BusinessException;
 
-import static com.ureca.snac.common.BaseCode.TOSS_NOT_ENOUGH_BALANCE;
+import static com.ureca.snac.common.BaseCode.GATEWAY_INSUFFICIENT_BALANCE;
 
 public class TossNotEnoughBalanceException extends BusinessException {
+
     public TossNotEnoughBalanceException() {
-        super(TOSS_NOT_ENOUGH_BALANCE);
+        super(GATEWAY_INSUFFICIENT_BALANCE);
     }
 }
