+++ src/main/java/com/ureca/snac/infra/exception/TossPaymentNotFoundException.java
@@ -1,0 +1,12 @@
+package com.ureca.snac.infra.exception;
+
+import com.ureca.snac.common.exception.ExternalApiException;
+
+import static com.ureca.snac.common.BaseCode.PAYMENT_NOT_FOUND;
+
+public class TossPaymentNotFoundException extends ExternalApiException {
+
+    public TossPaymentNotFoundException() {
+        super(PAYMENT_NOT_FOUND);
+    }
+}
